requirements.txt
Flask==2.0.2
python_dateutil==2.8.2
pytils==0.3
pytz==2021.3
requests==2.26.0
